requirement.txt
Python==3.7.0
certifi==2018.8.24
chardet==3.0.4
Django==2.0.8
django-rest-auth==0.9.3
django-static-precompiler==1.8.2
django-tastypie==0.14.2
djangorestframework==3.8.2
djangotoolbox==0.9.2
idna==2.7
Pillow==5.3.0
python-dateutil==2.7.3
python-mimeparse==1.6.0
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
whitenoise==4.1
django-cleanup=2.1.0
django_wysiwyg=0.8.0